added user's home page - catalog + sorting via criterias (+ validation!!!)

+++ src/ua/nure/stepanenko/SummaryTask4/db/entity/Entity.java
@@ -1,13 +1,13 @@
 package ua.nure.stepanenko.SummaryTask4.db.entity;
 
 public abstract class Entity {
-    protected long id;
+    protected int id;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
